Change BBText fontSize's type to number

+++ Extensions/BBText/bbtextruntimeobject.js
@@ -24,8 +24,9 @@
 gdjs.BBTextRuntimeObject = function(runtimeScene, objectData) {
   gdjs.RuntimeObject.call(this, runtimeScene, objectData);
 
+  // parseFloat should not be required, but GDevelop 5.0 beta 92 and below were storing it as a string.
   /** @type {number} */
-  this._opacity = parseInt(objectData.content.opacity);
+  this._opacity = parseFloat(objectData.content.opacity);
   /** @type {boolean} */
   this._visible = objectData.content.visible;
   /** @type {string} */
@@ -34,8 +35,9 @@ gdjs.BBTextRuntimeObject = function(runtimeScene, objectData) {
   this._color = objectData.content.color;
   /** @type {string} */
   this._fontFamily = objectData.content.fontFamily;
+  // parseFloat should not be required, but GDevelop 5.0 beta 92 and below were storing it as a string.
   /** @type {number} */
-  this._fontSize = parseInt(objectData.content.fontSize);
+  this._fontSize = parseFloat(objectData.content.fontSize);
   /** @type {boolean} */
   this._wordWrap = objectData.content.wordWrap;
   /** @type {number} */
